pdf_management_service/requirements_rag.txt
# Updated requirements for RAG Pipeline
Flask>=3.0.0
flask-cors>=4.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
firebase-admin>=6.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
Werkzeug>=3.0.0

# New dependencies for RAG pipeline
openai>=1.3.0
pinecone-client>=2.2.4
nltk>=3.8.1
tiktoken>=0.5.1
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional dependencies for enhanced functionality
requests>=2.31.0
pandas>=2.0.0
tqdm>=4.65.0
